Support PW entry retries for encrypted sshkey for sf_client
This adds support for password entry retries for signframework client, in case
the user fat-fingers the password entry (to unlock the encrypted sshkey).  It
only patches the sendFileToServer() function, since this seems to be the first
and only point where the user is actually prompted.  The number of retries is
configable by a defined macro in the code.

Signed-off-by: Dave Heller <hellerda@us.ibm.com>