Add RSA 2048/4096 "raw" signing support
Some existing hardware requires a custom DigestInfo/padding scheme.
Instead of supporting all of these, the signer can now pre-format
the padding on their hash and pass it as a raw payload to the
framework.

By doing a raw signing, it is up to the signer to verify that their
raw payload is correct before sending it to the server and verifying
that the signature can be successfully validated. The framework can
only decrypt the signature with the public key and compare it to the
input raw payload.