Fixed licensing, printing statements and makefile
This commit is in response to a second round of comments in the PR. I have fixed the license from proprietary to Apache 2.0. Additionally, I have responded to comments regarding printing type size_t. I also removed some commented out code in the sign_cert makefile and added the sign_cert makefile to be called from the main makefile.