Add new functionality for returning an x509 file (#17)
* Add new functionality for returning an x509 file

In this commit, I ported Ken Goldman's x509 generator code from 2013. This will allow for a client to request an x509 version of their public key. The x509 can be either self signed or signed from a CA, depending on priviledges. The system administrator can make a CA by generating a new key with an auxiliary file containing signature/digest algorithms, validity, final and Issuer fields. The significance of these fields can be found in Kens documentation. The key to be signed in the ceritficate can be a any project if the certifying key is declared as final. If not, then an auxiliary file will be needed for the key to be certified. This will contain all the same fields as the ceritifying aux file. If the certifying key is listed as final than this means that it is the final issuer in the certificate chain, it will be used to sign other projects only. The auxiliary file for these final issuers must also contain subject fields that describes the project it is signing. This is all described in kens documentation. To request an x509 from the server side, the user can run './sf_client ./sf_client -project <project name1> -comments "Testing sign cert" -param "-certify_project <project name2> -usr <userid>" -url sftp://<userid>@<framework host> -epwd <encrypted pswd> -pkey <rsa_private> ' where <project name1> is certifying/signing <project_name2> .
NOTE: You will likely need to increase the file_max field of your framework config, since x509s are larger than other outputed filed
This commit also comes with testing scripts. To run, follow the below steps:
1. run keygen commands described in generateTestFiles.sh
2. Add yourself to senders in the config files for issuer_final issuer_root code_signer
3. run generateTestFiles.sh
4. export FRAMEWORK_CONFIG_FILE to point to regression.cfg
5. run the framework
6. switch over to the client machine
7. fill in user specific fields in sign_cert_test.sh
8. run sign_cert_test.sh