Fix compiler warnings for openssl 3.0 and curl 7.85
The new compiler gcc-13 is flagging warnings in addition to openssl 3.0
and curl 7.85.
Fix the compiler warnings such as std::min requiring the same parameter
types.
Fix the openssl warnings by replacing the deprecated APIs if version is
3.0 or above.
Fix the curl warnings by replacing the deprecated APIs if version is
7.85 or above.

Tested-by: Matthew Vaught <matthew.vaught@ibm.com>

- RHEL8 with openssl 1.1 and libcurl 7.61.1
- MacOS with openssl 1.1, 3.0 and libcurl 7.87.0

Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>