client-c++: add generated file dependency
When building this in the bitbake env (which runs everything it can in
parallel), it hit an issue where the generated h file was not available
when trying to build the application that includes it.

Fix this issue by ensuring the application is not built until the header
file is generated.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>